+++ Packages/ai.fxn.fxn3d/Runtime/Internal/Function.cs
@@ -12,6 +12,6 @@ namespace Function.Internal {
 
     public static class Function {
 
-        public const string Version = @"0.0.3";
+        public const string Version = @"0.0.4";
     }
 }